[Agent] split mock factories
