+++ tests/common/mockFactories/index.js
@@ -1,4 +1,8 @@
 export * from './coreServices.js';
+export * from './loggerMocks.js';
+export * from './eventBusMocks.js';
+export * from './turnMocks.js';
+export * from './dataFetcherMock.js';
 export * from './loaders.js';
 export * from './entities.js';
 export * from './container.js';
@@ -6,11 +10,14 @@ export * from './container.js';
 // Explicit convenience exports
 export {
   createMockPathResolver,
-  createMockDataFetcher,
-  createMockValidatedEventDispatcherForIntegration,
-  createCapturingEventBus,
   createMockAIPromptPipeline,
 } from './coreServices.js';
+
+export { createMockDataFetcher } from './dataFetcherMock.js';
+export {
+  createMockValidatedEventDispatcherForIntegration,
+  createCapturingEventBus,
+} from './eventBusMocks.js';
 
 export { createLoaderMocks } from './loaders.js';
