+++ tests/common/mockFactories/loggerMocks.js
@@ -1,0 +1,4 @@
+import { createSimpleMock } from './coreServices.js';
+
+export const createMockLogger = () =>
+  createSimpleMock(['info', 'warn', 'error', 'debug']);
